scraper/requirements_GTS.txt
fastapi
uvicorn
pandas
mysql-connector-python
python-dotenv
feedparser
praw
pytrends
urllib3>=2.0
